+++ src/MicroElements.Metadata/Metadata/MutablePropertyContainer.cs
@@ -41,6 +41,7 @@ namespace MicroElements.Metadata
         /// <inheritdoc />
         public object GetValueUntyped(IProperty property, bool searchInParent = true)
         {
+            //TODO: COMPILE
             Type propertyType = property.Type;
             MethodInfo getValue = GetType().GetMethod(nameof(GetValue));
             MethodInfo getValueTyped = getValue.MakeGenericMethod(propertyType);
@@ -89,23 +90,41 @@ namespace MicroElements.Metadata
         /// <returns><see cref="IPropertyValue{T}"/> that holds value for property.</returns>
         public IPropertyValue<T> SetValue<T>(IProperty<T> property, T value)
         {
-            var newPropertyValue = new PropertyValue<T>(property, value, ValueSource.Defined);
+            var propertyValue = new PropertyValue<T>(property, value, ValueSource.Defined);
+            SetValue(propertyValue);
+            return propertyValue;
+        }
+
+        /// <inheritdoc />
+        public void SetValue(IPropertyValue propertyValue)
+        {
             bool isSet = false;
             for (int i = 0; i < _propertyValues.Count; i++)
             {
-                var propertyValue = _propertyValues[i];
-                if (propertyValue.PropertyUntyped == property)
+                var existingPropertyValue = _propertyValues[i];
+                if (existingPropertyValue.PropertyUntyped == propertyValue.PropertyUntyped)
                 {
-                    _propertyValues[i] = newPropertyValue;
+                    // replaces existing
+                    _propertyValues[i] = propertyValue;
                     isSet = true;
                     break;
                 }
             }
 
             if (!isSet)
-                _propertyValues.Add(newPropertyValue);
+                _propertyValues.Add(propertyValue);
+        }
 
-            return newPropertyValue;
+        /// <inheritdoc />
+        public void Add(IPropertyValue propertyValue)
+        {
+            _propertyValues.Add(propertyValue);
+        }
+
+        /// <inheritdoc />
+        public void Clear()
+        {
+            _propertyValues.Clear();
         }
 
         #endregion
@@ -122,7 +141,7 @@ namespace MicroElements.Metadata
         /// <param name="propertyContainer">MutablePropertyContainer.</param>
         /// <param name="parentPropertySource">Parent property source.</param>
         /// <returns>The same container with changed parent.</returns>
-        public static MutablePropertyContainer WithParentPropertySource(this MutablePropertyContainer propertyContainer, IPropertyContainer parentPropertySource)
+        public static IMutablePropertyContainer WithParentPropertySource(this IMutablePropertyContainer propertyContainer, IPropertyContainer parentPropertySource)
         {
             propertyContainer.SetParentPropertySource(parentPropertySource);
             return propertyContainer;
@@ -136,10 +155,107 @@ namespace MicroElements.Metadata
         /// <param name="property">Property to set.</param>
         /// <param name="value">Value to set.</param>
         /// <returns>The same container with changed property.</returns>
-        public static MutablePropertyContainer WithValue<T>(this MutablePropertyContainer propertyContainer, IProperty<T> property, T value)
+        public static IMutablePropertyContainer WithValue<T>(this IMutablePropertyContainer propertyContainer, IProperty<T> property, T value)
         {
             propertyContainer.SetValue(property, value);
             return propertyContainer;
+        }
+
+        /// <summary>
+        /// Sets value by string property name and returns the same container.
+        /// Overrides property value if exists with the same <paramref name="propertyName"/>.
+        /// </summary>
+        /// <typeparam name="T">Property type.</typeparam>
+        /// <param name="propertyContainer">MutablePropertyContainer.</param>
+        /// <param name="propertyName">Property name.</param>
+        /// <param name="value">Value to set.</param>
+        /// <returns>The same container with changed property.</returns>
+        public static IMutablePropertyContainer WithValue<T>(this IMutablePropertyContainer propertyContainer, string propertyName, T value)
+        {
+            propertyContainer.SetValue(propertyName, value);
+            return propertyContainer;
+        }
+
+        /// <summary>
+        /// Sets property value and returns the same container.
+        /// </summary>
+        /// <param name="propertyContainer">MutablePropertyContainer.</param>
+        /// <param name="property">Property to set.</param>
+        /// <param name="value">Value to set.</param>
+        /// <returns>The same container with changed property.</returns>
+        public static IMutablePropertyContainer WithValue(this IMutablePropertyContainer propertyContainer, IProperty property, object value)
+        {
+            propertyContainer.SetValue(property, value);
+            return propertyContainer;
+        }
+
+        /// <summary>
+        /// Sets value by string property name.
+        /// Overrides property value if exists with the same <paramref name="propertyName"/>.
+        /// </summary>
+        /// <typeparam name="T">Property type.</typeparam>
+        /// <param name="propertyContainer">MutablePropertyContainer.</param>
+        /// <param name="propertyName">Property name.</param>
+        /// <param name="value">Value to set.</param>
+        /// <returns><see cref="IPropertyValue{T}"/> that holds value for property.</returns>
+        public static IPropertyValue<T> SetValue<T>(this IMutablePropertyContainer propertyContainer, string propertyName, T value)
+        {
+            Type valueType = typeof(T);
+            IPropertyValue propertyValue = propertyContainer.GetPropertyValueUntyped(Search.ByNameOrAlias(propertyName).SearchInParent());
+
+            if (propertyValue != null)
+            {
+                IProperty existingProperty = propertyValue.PropertyUntyped;
+                if (existingProperty.Type != valueType)
+                {
+                    throw new ArgumentException($"Existing property {existingProperty.Name} has type {existingProperty.Type} but value has type {valueType}");
+                }
+
+                return propertyContainer.SetValue((IProperty<T>)existingProperty, value);
+            }
+
+            return propertyContainer.SetValue(new Property<T>(propertyName), value);
+        }
+
+        /// <summary>
+        /// Sets value by string property name.
+        /// Overrides property value if exists with the same <paramref name="propertyName"/>.
+        /// </summary>
+        /// <param name="propertyContainer">MutablePropertyContainer.</param>
+        /// <param name="propertyName">Property name.</param>
+        /// <param name="value">Value to set.</param>
+        /// <returns><see cref="IPropertyValue"/> that holds value for property.</returns>
+        public static IPropertyValue SetValue(this IMutablePropertyContainer propertyContainer, string propertyName, object value)
+        {
+            Type valueType = value.GetType();
+            IPropertyValue propertyValue = propertyContainer.GetPropertyValueUntyped(Search.ByNameOrAlias(propertyName).SearchInParent());
+
+            if (propertyValue != null)
+            {
+                IProperty existingProperty = propertyValue.PropertyUntyped;
+                if (existingProperty.Type != valueType)
+                {
+                    throw new ArgumentException($"Existing property {existingProperty.Name} has type {existingProperty.Type} but value has type {valueType}");
+                }
+
+                return propertyContainer.SetValue(existingProperty, value);
+            }
+
+            return propertyContainer.SetValue(Property.Create(valueType, propertyName), value);
+        }
+
+        /// <summary>
+        /// Sets property and value (non generic version).
+        /// </summary>
+        /// <param name="propertyContainer">PropertyContainer to change.</param>
+        /// <param name="property">Property to set.</param>
+        /// <param name="value">Value to set.</param>
+        /// <returns><see cref="IPropertyValue"/> that holds value for property.</returns>
+        public static IPropertyValue SetValue(this IMutablePropertyContainer propertyContainer, IProperty property, object value)
+        {
+            IPropertyValue propertyValue = PropertyValue.Create(property, value);
+            propertyContainer.SetValue(propertyValue);
+            return propertyValue;
         }
     }
 }