+++ src/MicroElements.Metadata/Utils/CodeCompiler.cs
@@ -2,68 +2,120 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using System;
+using System.Collections.Concurrent;
 using System.Linq.Expressions;
 using System.Reflection;
-using MicroElements.Functional;
 
 namespace MicroElements
 {
     public static class CodeCompiler
     {
+        #region Caches
+
+        static class Cache<Arg0, Result>
+        {
+            public static ConcurrentDictionary<Type, Func<Arg0, Result>> FuncCache = new ConcurrentDictionary<Type, Func<Arg0, Result>>();
+        }
+
+        static class Cache<Arg0, Arg1, Result>
+        {
+            public static ConcurrentDictionary<Type, Func<Arg0, Arg1, Result>> FuncCache = new ConcurrentDictionary<Type, Func<Arg0, Arg1, Result>>();
+        }
+
+        static class Cache<Arg0, Arg1, Arg2, Result>
+        {
+            public static ConcurrentDictionary<Type, Func<Arg0, Arg1, Arg2, Result>> FuncCache = new ConcurrentDictionary<Type, Func<Arg0, Arg1, Arg2, Result>>();
+        }
+
+        #endregion
 
         /// <summary>
         /// Маркерный тип для указания, что функция будет сгенерирована для произвольных типов.
         /// </summary>
         public class GenericType { }
 
-        public static Func<Type, Func<Arg0, Result>> CreateCompiledFunc<Arg0, Result>(Func<Arg0, Result> methodFunc)
+        public static Func<Arg0, Result> CachedCompiledFunc<Arg0, Result>(Type type, Func<Arg0, Result> genericMethodFunc)
         {
-            MethodInfo methodInfo = methodFunc.Method.GetGenericMethodDefinition();
-            return CreateCompiledFunc<Arg0, Result>(methodInfo);
+            if (!Cache<Arg0, Result>.FuncCache.TryGetValue(type, out var cachedFunc))
+            {
+                cachedFunc = CompileGeneric(type, genericMethodFunc);
+                Cache<Arg0, Result>.FuncCache.TryAdd(type, cachedFunc);
+            }
+
+            return cachedFunc;
         }
 
-        public static Func<Type, Func<Arg0, Arg1, Result>> CreateCompiledFunc<Arg0, Arg1, Result>(Func<Arg0, Arg1, Result> methodFunc)
+        public static Func<Arg0, Arg1, Result> CachedCompiledFunc<Arg0, Arg1, Result>(Type type, Func<Arg0, Arg1, Result> genericMethodFunc)
         {
-            MethodInfo methodInfo = methodFunc.Method.GetGenericMethodDefinition();
-            return CreateCompiledFunc<Arg0, Arg1, Result>(methodInfo);
+            if (!Cache<Arg0, Arg1, Result>.FuncCache.TryGetValue(type, out var cachedFunc))
+            {
+                cachedFunc = CompileGeneric(type, genericMethodFunc);
+                Cache<Arg0, Arg1, Result>.FuncCache.TryAdd(type, cachedFunc);
+            }
+
+            return cachedFunc;
         }
 
-        public static Func<Type, Func<Arg0, Result>> CreateCompiledFunc<Arg0, Result>(Type methodOwnerType, string methodName)
+        public static Func<Arg0, Arg1, Arg2, Result> CachedCompiledFunc<Arg0, Arg1, Arg2, Result>(Type type, Func<Arg0, Arg1, Arg2, Result> genericMethodFunc)
         {
-            MethodInfo methodInfo = methodOwnerType.GetMethod(methodName, BindingFlags.Public | BindingFlags.Static);
-            return CreateCompiledFunc<Arg0, Result>(methodInfo);
+            if (!Cache<Arg0, Arg1, Arg2, Result>.FuncCache.TryGetValue(type, out var cachedFunc))
+            {
+                cachedFunc = CompileGeneric(type, genericMethodFunc);
+                Cache<Arg0, Arg1, Arg2, Result>.FuncCache.TryAdd(type, cachedFunc);
+            }
+
+            return cachedFunc;
         }
 
-        public static Func<Type, Func<Arg0, Arg1, Result>> CreateCompiledFunc<Arg0, Arg1, Result>(Type methodOwnerType, string methodName)
+        private static Func<Arg0, Result> CompileGeneric<Arg0, Result>(Type propertyType, Func<Arg0, Result> genericMethodFunc)
         {
-            MethodInfo methodInfo = methodOwnerType.GetMethod(methodName, BindingFlags.Public | BindingFlags.Static);
-            return CreateCompiledFunc<Arg0, Arg1, Result>(methodInfo);
+            MethodInfo methodInfo = genericMethodFunc.Method.GetGenericMethodDefinition();
+            return Compile<Arg0, Result>(propertyType, methodInfo);
         }
 
-        public static Func<Type, Func<Arg0, Result>> CreateCompiledFunc<Arg0, Result>(MethodInfo methodInfo, int cacheLimit = 20) =>
-            Prelude.Memoize<Type, Func<Arg0, Result>>(
-                propertyType =>
-                {
-                    MethodInfo genericMethod = methodInfo.MakeGenericMethod(propertyType);
-                    ParameterExpression arg0 = Expression.Parameter(typeof(Arg0));
-                    MethodCallExpression callExpression = Expression.Call(genericMethod, arg0);
-                    return Expression
-                        .Lambda<Func<Arg0, Result>>(callExpression, arg0)
-                        .Compile();
-                }, cacheLimit);
+        private static Func<Arg0, Arg1, Result> CompileGeneric<Arg0, Arg1, Result>(Type propertyType, Func<Arg0, Arg1, Result> genericMethodFunc)
+        {
+            MethodInfo methodInfo = genericMethodFunc.Method.GetGenericMethodDefinition();
+            return Compile<Arg0, Arg1, Result>(propertyType, methodInfo);
+        }
 
-        public static Func<Type, Func<Arg0, Arg1, Result>> CreateCompiledFunc<Arg0, Arg1, Result>(MethodInfo methodInfo, int cacheLimit = 20) =>
-            Prelude.Memoize<Type, Func<Arg0, Arg1, Result>>(
-                propertyType =>
-                {
-                    MethodInfo genericMethod = methodInfo.MakeGenericMethod(propertyType);
-                    ParameterExpression arg0 = Expression.Parameter(typeof(Arg0));
-                    ParameterExpression arg1 = Expression.Parameter(typeof(Arg1));
-                    MethodCallExpression callExpression = Expression.Call(genericMethod, arg0, arg1);
-                    return Expression
-                        .Lambda<Func<Arg0, Arg1, Result>>(callExpression, arg0, arg1)
-                        .Compile();
-                }, cacheLimit);
+        private static Func<Arg0, Arg1, Arg2, Result> CompileGeneric<Arg0, Arg1, Arg2, Result>(Type propertyType, Func<Arg0, Arg1, Arg2, Result> genericMethodFunc)
+        {
+            MethodInfo methodInfo = genericMethodFunc.Method.GetGenericMethodDefinition();
+            return Compile<Arg0, Arg1, Arg2, Result>(propertyType, methodInfo);
+        }
 
+        private static Func<Arg0, Result> Compile<Arg0, Result>(Type propertyType, MethodInfo methodInfo)
+        {
+            MethodInfo genericMethod = methodInfo.MakeGenericMethod(propertyType);
+            ParameterExpression arg0 = Expression.Parameter(typeof(Arg0));
+            MethodCallExpression callExpression = Expression.Call(genericMethod, arg0);
+            return Expression
+                .Lambda<Func<Arg0, Result>>(callExpression, arg0)
+                .Compile();
+        }
+
+        private static Func<Arg0, Arg1, Result> Compile<Arg0, Arg1, Result>(Type propertyType, MethodInfo methodInfo)
+        {
+            MethodInfo genericMethod = methodInfo.MakeGenericMethod(propertyType);
+            ParameterExpression arg0 = Expression.Parameter(typeof(Arg0));
+            ParameterExpression arg1 = Expression.Parameter(typeof(Arg1));
+            MethodCallExpression callExpression = Expression.Call(genericMethod, arg0, arg1);
+            return Expression
+                .Lambda<Func<Arg0, Arg1, Result>>(callExpression, arg0, arg1)
+                .Compile();
+        }
+
+        private static Func<Arg0, Arg1, Arg2, Result> Compile<Arg0, Arg1, Arg2, Result>(Type propertyType, MethodInfo methodInfo)
+        {
+            MethodInfo genericMethod = methodInfo.MakeGenericMethod(propertyType);
+            ParameterExpression arg0 = Expression.Parameter(typeof(Arg0));
+            ParameterExpression arg1 = Expression.Parameter(typeof(Arg1));
+            ParameterExpression arg2 = Expression.Parameter(typeof(Arg2));
+            MethodCallExpression callExpression = Expression.Call(genericMethod, arg0, arg1, arg2);
+            return Expression
+                .Lambda<Func<Arg0, Arg1, Arg2, Result>>(callExpression, arg0, arg1, arg2)
+                .Compile();
+        }
     }
 }